requirements.txt
MetaTrader5
python-dotenv
pandas
numpy
scikit-learn
tensorflow   # or keras
matplotlib
requests 